requirements.txt
python-telegram-bot[rate-limiter,job-queue]>=21.4,<22
pydantic>=2.6,<3
pydantic-settings>=2.2,<3
uvloop; sys_platform != "win32"
httpx>=0.27
aiosqlite>=0.20
